Add Bootstrap menu and style transactions
