+++ app/static/css/layout.css
@@ -1,0 +1,7 @@
+body {
+  font-family: sans-serif;
+}
+
+#table-container {
+  overflow-y: auto;
+}
