+++ app/static/index.html
@@ -1,53 +1,86 @@
 <!DOCTYPE html>
 <html lang="es">
 <head>
-    <meta charset="UTF-8" />
-    <title>Movimientos</title>
-    <style>
-        body { font-family: sans-serif; margin:0; padding:0; display:flex; flex-direction:column; height:100vh; }
-        header, footer { background:#f0f0f0; padding:0.5rem; }
-        #table-container { flex:1; overflow-y:auto; }
-        table { width:100%; border-collapse:collapse; }
-        th, td { padding:0.25rem; border-bottom:1px solid #ccc; }
-        #modal { display:none; position:fixed; top:0; left:0; width:100%; height:100%; background:rgba(0,0,0,0.4); justify-content:center; align-items:center; }
-        #modal form { background:white; padding:1rem; }
-    </style>
+  <meta charset="UTF-8" />
+  <title>Movimientos</title>
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
+  <link rel="stylesheet" href="css/layout.css">
 </head>
-<body>
-    <header>
-        <button id="config-btn">Configuración</button>
-    </header>
-    <div id="table-container">
-        <table id="tx-table">
-            <thead>
-                <tr>
-                    <th>Fecha</th>
-                    <th>Concepto</th>
-                    <th>Monto</th>
-                    <th>Tipo</th>
-                    <th>Cuenta</th>
-                </tr>
-            </thead>
-            <tbody></tbody>
-        </table>
+<body class="d-flex flex-column min-vh-100">
+  <header class="navbar navbar-light bg-light p-2">
+    <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas" data-bs-target="#sideMenu" aria-controls="sideMenu">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="d-flex gap-3 ms-auto">
+      <button id="add-income" class="btn btn-success">+ Ingreso</button>
+      <button id="add-expense" class="btn btn-danger">+ Egreso</button>
     </div>
-    <footer>
-        <button id="add-income">+ Ingreso</button>
-        <button id="add-expense">+ Egreso</button>
-    </footer>
-
-    <div id="modal">
+  </header>
+  <div class="offcanvas offcanvas-start" tabindex="-1" id="sideMenu" aria-labelledby="sideMenuLabel">
+    <div class="offcanvas-header">
+      <h5 id="sideMenuLabel" class="offcanvas-title">Menú</h5>
+      <button type="button" class="btn-close" data-bs-dismiss="offcanvas"></button>
+    </div>
+    <div class="offcanvas-body">
+      <ul class="list-unstyled">
+        <li><a href="/config" class="text-decoration-none">Configuración</a></li>
+      </ul>
+    </div>
+  </div>
+  <div id="table-container" class="flex-grow-1 overflow-auto">
+    <table id="tx-table" class="table table-striped mb-0">
+      <thead>
+        <tr>
+          <th>Fecha</th>
+          <th>Concepto</th>
+          <th>Monto</th>
+          <th>Tipo</th>
+          <th>Cuenta</th>
+        </tr>
+      </thead>
+      <tbody></tbody>
+    </table>
+  </div>
+  <div class="modal fade" id="txModal" tabindex="-1">
+    <div class="modal-dialog">
+      <div class="modal-content">
         <form id="tx-form">
-            <h3 id="form-title"></h3>
-            <label>Fecha <input type="date" name="date" required></label><br/>
-            <label>Concepto <input type="text" name="description"></label><br/>
-            <label>Monto <input type="number" step="0.01" name="amount" required></label><br/>
-            <label>Cuenta <select name="account_id"></select></label><br/>
-            <button type="submit">Guardar</button>
-            <button type="button" id="cancel-btn">Cancelar</button>
+          <div class="modal-header">
+            <h5 class="modal-title" id="form-title"></h5>
+            <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
+          </div>
+          <div class="modal-body">
+            <div class="mb-3">
+              <label class="form-label">Fecha
+                <input type="date" name="date" class="form-control" required>
+              </label>
+            </div>
+            <div class="mb-3">
+              <label class="form-label">Concepto
+                <input type="text" name="description" class="form-control">
+              </label>
+            </div>
+            <div class="mb-3">
+              <label class="form-label">Monto
+                <input type="number" step="0.01" name="amount" class="form-control" required>
+              </label>
+            </div>
+            <div class="mb-3">
+              <label class="form-label">Cuenta
+                <select name="account_id" class="form-select"></select>
+              </label>
+            </div>
+          </div>
+          <div class="modal-footer">
+            <button type="submit" class="btn btn-primary">Guardar</button>
+            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancelar</button>
+          </div>
         </form>
+      </div>
     </div>
+  </div>
 
-    <script src="main.js"></script>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
