+++ app/static/js/api.js
@@ -1,0 +1,17 @@
+export async function fetchAccounts() {
+  const res = await fetch('/accounts');
+  return res.json();
+}
+
+export async function fetchTransactions(limit, offset) {
+  const res = await fetch(`/transactions?limit=${limit}&offset=${offset}`);
+  return res.json();
+}
+
+export async function createTransaction(payload) {
+  await fetch('/transactions', {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(payload)
+  });
+}
